+++ sdk/dotnet/Provider.cs
@@ -53,6 +53,16 @@ namespace Pulumi.Kubernetes
         public Input<string>? Cluster { get; set; }
 
         /// <summary>
+        /// If present, this value will control the provider's replacement behavior. In particular, the provider will _only_ be replaced when `clusterIdentifier` changes; all other changes to provider configuration will be treated as updates.
+        /// 
+        /// Kubernetes does not yet offer an API for cluster identification, so Pulumi uses heuristics to decide when a provider resource should be replaced or updated. These heuristics can sometimes lead to destructive replace operations when an update would be more appropriate, or vice versa.
+        /// 
+        /// Use `clusterIdentifier` for more fine-grained control of the provider resource's lifecycle.
+        /// </summary>
+        [Input("clusterIdentifier")]
+        public Input<string>? ClusterIdentifier { get; set; }
+
+        /// <summary>
         /// If present, the name of the kubeconfig context to use.
         /// </summary>
         [Input("context")]