+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### Added
 
+- `clusterIdentifier` configuration can now be used to manually control the
+  replacement behavior of a provider resource.
+  (https://github.com/pulumi/pulumi-kubernetes/pull/3068)
+
 - Pod errors now include the pod's last termination state, as well as the pod's
   termination message if available.
   (https://github.com/pulumi/pulumi-kubernetes/pull/3091)
@@ -16,11 +20,12 @@
   Use `terminationMessagePolicy: FallbackToLogsOnError` to use the pod's logs
   as its termination message.
 
-- Define Supported Languages for Overlays (https://github.com/pulumi/pulumi-kubernetes/pull/3107)
+- Documentation is now generated for all languages supported by overlay types.
+  (https://github.com/pulumi/pulumi-kubernetes/pull/3107)
 
 ### Fixed
 
-- Updated logic to accurately detect if a resource is a Patch variant (https://github.com/pulumi/pulumi-kubernetes/pull/3102)
+- Updated logic to accurately detect if a resource is a Patch variant. (https://github.com/pulumi/pulumi-kubernetes/pull/3102)
 
 ## 4.15.0 (July 9, 2024)